グラフデータの取得・管理といったロジックはカスタムフックに分離したい
