+++ src/features/populationChart/index.tsx
@@ -10,58 +10,23 @@ import {
   XAxis,
   YAxis,
 } from "recharts";
-import { PopulationLabel, PopulationResponse, Prefecture } from "@/types";
-import { useEffect, useState } from "react";
+import { Prefecture } from "@/types";
 import { useSelectedPrefecturesStore } from "@/stores/selectedPrefectures";
-import { populationLabels } from "@/constants";
+import { usePopulationData } from "@/hooks/usePopulationData";
 import styles from "./index.module.css";
+import { populationLabels } from "@/constants";
 
 type Props = {
   prefectures: Prefecture[];
 };
 
-type PopulationChartData = {
-  year: number;
-  [key: string]: number;
-};
-
 export default function PopulationChart({ prefectures }: Props) {
   const { selectedPrefectures } = useSelectedPrefecturesStore();
-  const [data, setData] = useState<PopulationChartData[]>([]);
-  const [selectedPopulationLabel, setSelectedPopulationLabel] =
-    useState<PopulationLabel>("総人口");
-
-  useEffect(() => {
-    const fetchPopulationData = async () => {
-      const promises = selectedPrefectures.map((prefCode) =>
-        fetchPopulation(prefCode)
-      );
-      const populationDataArray = await Promise.all(promises);
-      const newData: PopulationChartData[] = [];
-
-      populationDataArray.forEach((prefectureData, index) => {
-        const prefCode = selectedPrefectures[index];
-        const prefName = prefectures.find(
-          (p) => p.prefCode === prefCode
-        )?.prefName;
-
-        prefectureData
-          .find(({ label }) => label === selectedPopulationLabel)
-          ?.data.forEach((item) => {
-            const existingItem = newData.find((d) => d.year === item.year);
-            if (existingItem) {
-              existingItem[prefName!] = item.value;
-            } else {
-              newData.push({ year: item.year, [prefName!]: item.value });
-            }
-          });
-      });
-
-      setData(newData);
-    };
-
-    fetchPopulationData();
-  }, [selectedPrefectures, prefectures, selectedPopulationLabel]);
+  const {
+    populationData,
+    selectedPopulationLabel,
+    setSelectedPopulationLabel,
+  } = usePopulationData(selectedPrefectures, prefectures);
 
   return (
     <>
@@ -81,7 +46,7 @@ export default function PopulationChart({ prefectures }: Props) {
           ))}
         </div>
         <ResponsiveContainer width="100%" height={500}>
-          <LineChart data={data} className={styles.lineChart}>
+          <LineChart data={populationData} className={styles.lineChart}>
             <CartesianGrid strokeDasharray="3 3" />
             <XAxis
               dataKey="year"
@@ -115,7 +80,7 @@ export default function PopulationChart({ prefectures }: Props) {
                 overflowY: "auto",
               }}
             />
-            {Object.keys(data[0] || {})
+            {Object.keys(populationData[0] || {})
               .filter((key) => key !== "year")
               .map((prefName, index) => {
                 const hue = (index * 137.5) % 360;
@@ -137,15 +102,4 @@ export default function PopulationChart({ prefectures }: Props) {
       </div>
     </>
   );
-}
-
-async function fetchPopulation(prefCode: number): Promise<PopulationResponse> {
-  try {
-    const res = await fetch(`/api/prefectures/${prefCode}/population`);
-    const data = await res.json();
-    return data;
-  } catch (error) {
-    console.error(error);
-    throw new Error("Failed to fetch population data");
-  }
 }
